src/workflows/batch_warnings_processor_v2/tests/requirements.txt
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
pytest-html>=3.1.0
pytest-json-report>=1.5.0
pytest-benchmark>=4.0.0
pytest-xdist>=3.0.0
aiohttp>=3.8.0
psutil>=5.9.0
openai>=1.0.0
python-dotenv
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation

